refactor: renamed TouchableRipple file to sort docs properlt
