+++ src/components/Checkbox/CheckboxAndroid.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Animated, View, StyleSheet } from 'react-native';
 import color from 'color';
 import MaterialCommunityIcon from '../MaterialCommunityIcon';
-import TouchableRipple from '../TouchableRipple';
+import TouchableRipple from '../TouchableRipple/TouchableRipple';
 import { withTheme } from '../../core/theming';
 import type { $RemoveChildren } from '../../types';
