+++ src/components/List/ListAccordion.tsx
@@ -8,7 +8,7 @@ import {
   TextStyle,
   I18nManager,
 } from 'react-native';
-import TouchableRipple from '../TouchableRipple';
+import TouchableRipple from '../TouchableRipple/TouchableRipple';
 import MaterialCommunityIcon from '../MaterialCommunityIcon';
 import Text from '../Typography/Text';
 import { withTheme } from '../../core/theming';
